# nao_backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:5001",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/signin": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get all user already created",
        "parameters": [],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/AllUser"
            },
            "description": "User created succefully"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No user exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "No user exists"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "Create an user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/User"
            },
            "description": "User created succefully"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User already exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "User already exists"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/user/{name}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get user by name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/UserByName"
            },
            "description": "User created succefully"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User does not exist"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "User does not exist"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Get user by name",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "Jhon Doe"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Login Successfully"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User does not exist"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "User does not exist"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "userEmail": {
          "type": "string",
          "example": "JhonDoe@epitech.eu"
        }
      }
    },
...
    "AllUser": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              },
              "username": {
                "type": "string",
                "example": "Jhon Doe"
              },
              "email": {
                "type": "string",
                "example": "JhonDoe@epitech.eu"
              },
              "password": {
                "type": "string",
                "example": "password"
              }
            }
          }
        }
      },
      "required": [
        "result"
      ]
    },
    "UserByName": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              },
              "username": {
                "type": "string",
                "example": "Jhon Doe"
              },
              "email": {
                "type": "string",
                "example": "JhonDoe@epitech.eu"
              },
              "password": {
                "type": "string",
                "example": "password"
              }
            }
          }
        }
      },
      "required": [
        "result"
      ]
    },
...
  }
}